+++ routes/usersRoute.js
@@ -2,7 +2,7 @@ import express from "express";
 import User from "../models/userModel.js";
 import Post from "../models/postModel.js";
 const router = express.Router();
-import { comparePassword, hashPassword } from "../helpers/authHelper.js";
+import { hashPassword, controlAuthorization } from "../helpers/authHelper.js";
 
 //GET all users
 router.get("/", async (req, res) => {
@@ -44,29 +44,38 @@ router.get("/:id", async (req, res) => {
 });
 
 //Patch the user
-router.patch("/:id", async (req, res) => {
+router.patch("/:id", controlAuthorization, async (req, res) => {
     try {
         const { id } = req.params;
         const { password } = req.body;
-        //hash nuova password
         const user = await User.findById(id);
-        if (password !== user.password) {
-            const hashedPassword = await hashPassword(password);
-            req.body.password = hashedPassword;
+
+        let userUpdate;
+
+        if (req.user._id === req.params.id) {
+            //hash new password
+            if (password !== undefined && password !== user.password) {
+                const hashedPassword = await hashPassword(password);
+                req.body.password = hashedPassword;
+            }
+
+            userUpdate = await User.findByIdAndUpdate({ _id: id }, req.body, {
+                new: true,
+                runValidators: true,
+            }).populate({
+                path: "city",
+                select: "name",
+            });
+        } else {
+            return res.status(403).send(`Request not authorized`);
         }
 
-        const userUpdate = await User.findByIdAndUpdate({ _id: id }, req.body, {
-            new: true,
-            runValidators: true,
-        }).populate({
-            path: "city",
-            select: "name",
-        });
         if (!user) {
             return res
                 .status(404)
-                .json({ message: `Cant find user with ID ${id}` });
+                .json({ message: `Can't find user with ID ${id}` });
         }
+
         return res.status(200).json(userUpdate);
     } catch (error) {
         console.error(error.stack);
@@ -76,18 +85,21 @@ router.patch("/:id", async (req, res) => {
 
 //DELETE the user
 //Delete all the posts as well
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", controlAuthorization, async (req, res) => {
     try {
         const { id } = req.params;
         const user = await User.findById(id);
-        try {
+
+        console.log(req.user._id);
+        console.log(req.params.id);
+        if (req.user._id === req.params.id) {
             await Post.deleteMany({ user: user._id });
             await User.findByIdAndDelete(id);
             return res.status(200).json({
                 message: `The user ID${id} was erased from database`,
             });
-        } catch (error) {
-            res.status(500).json({ message: error.message });
+        } else {
+            return res.status(403).send(`Request not authorized`);
         }
     } catch (error) {
         console.error(error.stack);
